tsm/requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
amqp==1.4.9
anyio==3.6.2
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.1.0
beautifulsoup4==4.10.0
billiard==3.3.0.23
celery==3.1.18
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
coloredlogs==15.0.1
cryptography==37.0.4
cssselect2==0.6.0
Django==3.2.11
django-bootstrap4==21.2
django-bootstrap5==22.1
django-cleanup==6.0.0
django-cors-headers==3.12.0
django-phonenumber-field==6.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-suit==2.0a1
djangorestframework==3.13.1
easy-thumbnails==2.8.1
et-xmlfile==1.1.0
fastapi==0.87.0
filetype==1.2.0
Flask==1.1.2
flatbuffers==22.11.23
frozenlist==1.3.3
gunicorn==20.0.4
h11==0.13.0
humanfriendly==10.0
idna==3.3
ImageHash==4.3.1
imageio==2.22.4
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==3.0.37
llvmlite==0.39.1
lxml==4.8.0
MarkupSafe==1.1.1
mpmath==1.2.1
multidict==6.0.3
networkx==2.8.8
numba==0.56.4
numpy==1.23.5
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.2
phonenumberslite==8.12.51
Pillow==9.3.0
pooch==1.6.0
protobuf==4.21.10
pycparser==2.21
pydantic==1.10.2
PyMatting==1.1.8
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2020.1
PyWavelets==1.4.1
rembg==2.0.29
reportlab==3.6.9
requests==2.27.1
scikit-image==0.19.3
scipy==1.9.3
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.3.1
starlette==0.21.0
svglib==1.2.1
sympy==1.11.1
tifffile==2022.10.10
tinycss2==1.1.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
typing-extensions==4.4.0
urllib3==1.26.8
uvicorn==0.20.0
watchdog==2.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wsproto==1.1.0
yarl==1.8.2
